Do not cast language weight to int32!
git-svn-id: https://cmusphinx.svn.sourceforge.net/svnroot/cmusphinx/trunk@9113 c60273ff-4a12-0410-8d58-82ceea6d4170

+++ sphinx3/src/libs3decoder/liblm/lm.c
@@ -2142,9 +2142,8 @@ swap_tg32(tg32_t * t32)
 int32
 lm_rawscore(lm_t * lm, int32 score)
 {
-
     score -= lm->wip;
-    score /= (int32) lm->lw;
+    score /= lm->lw;
 
     return score;
 }
